made a manual function that checks if the user is a admin or not, and redirects them when they aren't

+++ routes/web.php
@@ -54,22 +54,22 @@ Route::post('/update-cart-item', [CartController::class, "updateCartItem"])->nam
 Route::get('/promo/create', [PromotionController::class, "createForm"])->middleware(['auth', 'verified'])->name('promo.createForm');
 Route::post('/promo/create/code', [PromotionController::class, "create"])->middleware(['auth', 'verified'])->name('promo.create');
 Route::get('/promo', [PromotionController::class, "show"])->middleware(['auth', 'verified'])->name('promo.show');
-Route::delete('/promo/delete/{id}', [PromotionController::class, "delete"])->name('promo.delete');
-Route::get('/promo/{id}/edit', [PromotionController::class, 'editForm'])->name('promo.edit');
-Route::put('/promo/{id}/update', [PromotionController::class, 'update'])->name('promo.update');
+Route::delete('/promo/delete/{id}', [PromotionController::class, "delete"])->middleware(['auth', 'verified'])->name('promo.delete');
+Route::get('/promo/{id}/edit', [PromotionController::class, 'editForm'])->middleware(['auth', 'verified'])->name('promo.edit');
+Route::put('/promo/{id}/update', [PromotionController::class, 'update'])->middleware(['auth', 'verified'])->name('promo.update');
 
 Route::get('/users', [ProfileController::class, "allUserDashboard"])->middleware(['auth', 'verified'])->name('users.show');
-Route::get('/users/create', [ProfileController::class, 'showCreate'])->name('users.showCreate');
-Route::get('/users/{id}/edit', [ProfileController::class, 'editForm'])->name('users.edit');
-Route::put('/users/{id}/update', [ProfileController::class, 'updateUser'])->name('users.update');
-Route::delete('/users/{id}', [ProfileController::class, 'delete'])->name('users.delete');
+Route::get('/users/create', [ProfileController::class, 'showCreate'])->middleware(['auth', 'verified'])->name('users.showCreate');
+Route::get('/users/{id}/edit', [ProfileController::class, 'editForm'])->middleware(['auth', 'verified'])->name('users.edit');
+Route::put('/users/{id}/update', [ProfileController::class, 'updateUser'])->middleware(['auth', 'verified'])->name('users.update');
+Route::delete('/users/{id}', [ProfileController::class, 'delete'])->middleware(['auth', 'verified'])->name('users.delete');
 
 Route::get('/productManagement', [ProductController::class, "allProductsDashboard"])->middleware(['auth', 'verified'])->name('products.show');
-Route::get('/productManagement/create', [ProductController::class, 'showCreate'])->name('products.showCreate');
-Route::post('/productManagement/create/submit', [ProductController::class, 'create'])->name('products.create');
-Route::get('/productManagement/{id}/edit', [ProductController::class, 'editForm'])->name('products.edit');
-Route::put('/productManagement/{id}/update', [ProductController::class, 'updateProduct'])->name('products.update');
-Route::delete('/productManagement/{id}', [ProductController::class, 'delete'])->name('products.delete');
+Route::get('/productManagement/create', [ProductController::class, 'showCreate'])->middleware(['auth', 'verified'])->name('products.showCreate');
+Route::post('/productManagement/create/submit', [ProductController::class, 'create'])->middleware(['auth', 'verified'])->name('products.create');
+Route::get('/productManagement/{id}/edit', [ProductController::class, 'editForm'])->middleware(['auth', 'verified'])->name('products.edit');
+Route::put('/productManagement/{id}/update', [ProductController::class, 'updateProduct'])->middleware(['auth', 'verified'])->name('products.update');
+Route::delete('/productManagement/{id}', [ProductController::class, 'delete'])->middleware(['auth', 'verified'])->name('products.delete');
 
 
 Route::middleware('auth')->group(function () {
